f - NPE on Mob spawn

+++ src/com/modcrafting/diablodrops/builders/TierBuilder.java
@@ -62,8 +62,8 @@ public class TierBuilder
             plugin.tiers.add(new Tier(name, ChatColor.valueOf(color
                     .toUpperCase()), Math.abs(amt), Math.abs(lvl), Math
                     .abs((int) (chance * 100)), l, lore, cs.getString(name
-                    + ".DisplayName", name), (float) cs.getDouble(name
-                    + ".DropChance", 100) / 100));
+                    + ".DisplayName", name), ((float) cs.getInt(name
+                    + ".DropChance", 100)) / 100));
         }
     }
 }
